Requests every countries' URLs
Requests the start_urls for every tracked country and formatted them in readable_output.ipynb

Ukraine is a special case, its inclusion is explained in readable_output.ipynb

I also chose to include null values for keys that I previously excluded (so as to better fit the given format on GitHub and make it simpler). This was especially useful since Iran (and possibly others) didn't fit the normal format.